+++ app/scripts/components/exploration/index.tsx
@@ -1,10 +1,4 @@
-import React, {
-  useEffect,
-  useState,
-  createContext,
-  useContext,
-  useCallback
-} from 'react';
+import React, { useEffect, useState } from 'react';
 import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
 import styled from 'styled-components';
 import { themeVal } from '@devseed-ui/theme-provider';
@@ -15,20 +9,8 @@ import { ExplorationMap } from './components/map';
 import { useAnalysisController } from './hooks/use-analysis-data-request';
 import { TimelineDataset } from './types.d.ts';
 import { selectedCompareDateAtom, selectedDateAtom } from './atoms/dates';
-import DataLayerCard from './components/datasets/data-layer-card'; // DataLayerCardWithSync
-// import useTimelineDatasetsAtom from './hooks/use-timeline-dataset-atom';
-import useTimelineDatasetAtom from './hooks/use-timeline-dataset-atom';
-import { externalDatasetsAtom } from './atoms/datasetLayers';
 import { CLEAR_LOCATION, urlAtom } from '$utils/params-location-atom/url';
 import { legacyGlobalStyleCSSBlock } from '$styles/legacy-global-styles';
-import { SetState } from '$types/aliases';
-import { DatasetData } from '$types/veda';
-import {
-  // useTimelineDatasetAtom,
-  useTimelineDatasetColormap,
-  useTimelineDatasetVisibility,
-  useTimelineDatasetColormapScale
-} from '$components/exploration/atoms/hooks';
 
 // @TODO: "height: 100%" Added for exploration container to show correctly in NextJs instance but investigate why this is needed and possibly work to remove
 const Container = styled.div`
@@ -141,167 +123,3 @@ export default function ExplorationAndAnalysis(
     </Container>
   );
 }
-
-// @NOTE-SANDRA: hooks only make sense to be used and imported when used within a system of E&A... maybe... this is what I need to explore
-// @NOTE-SANDRA: Lots of dependencies on the datasetAtom - useAnalysisDataRequest for example and it also needs the vedaUIProvider...
-// @NOTE-SANDRA: Hooks are attached to the urlParams... system as a whole depends on URL Params
-
-interface EACompoundContextType {
-  selectedDay: Date | null;
-  setSelectedDay: (day: Date) => void;
-  selectedCompareDay: Date | null;
-  setSelectedCompareDay: (day: Date) => void;
-  datasets: TimelineDataset[];
-  setDatasets: (datasets: TimelineDataset[]) => void;
-}
-
-interface EACompoundStateValues {
-  eaDatasets: TimelineDataset[];
-  setEaDatasets: (datasets: TimelineDataset[]) => void;
-  selectedDay: Date | null;
-  setSelectedDay: (day: Date) => void;
-  selectedCompareDay: Date | null;
-  setSelectedCompareDay: (day: Date) => void;
-  setEAExternalDatasets: SetState<DatasetData[]>;
-}
-
-export function useEACompoundState(): EACompoundStateValues {
-  const [timelineDatasets, setTimelineDatasets] = useTimelineDatasetAtom();
-  const [selectedDay, setSelectedDay] = useAtom(selectedDateAtom);
-  const [selectedCompareDay, setSelectedCompareDay] = useAtom(
-    selectedCompareDateAtom
-  );
-  const setEAExternalDatasets = useSetAtom(externalDatasetsAtom);
-
-  return {
-    eaDatasets: timelineDatasets,
-    setEaDatasets: setTimelineDatasets,
-    selectedDay,
-    setSelectedDay,
-    selectedCompareDay,
-    setSelectedCompareDay,
-    setEAExternalDatasets
-  };
-}
-
-const EACompoundContext = createContext<EACompoundContextType>({
-  selectedDay: null,
-  setSelectedDay: (day: Date) => {},
-  selectedCompareDay: null,
-  setSelectedCompareDay: (day: Date) => {},
-  datasets: [],
-  setDatasets: (datasets: TimelineDataset[]) => {}
-});
-
-const useEACompoundContext = () => {
-  const context = useContext(EACompoundContext);
-  if (!context) {
-    throw new Error(
-      'useEACompoundContext must be used within a ExplorationAndAnalysisCompound'
-    );
-  }
-  return context;
-};
-// http://localhost:9000/sandbox?datasets=%5B%7B%22id%22%3A%22casa-gfed-co2-flux-hr%22%2C%22settings%22%3A%7B%22isVisible%22%3Atrue%2C%22opacity%22%3A100%2C%22analysisMetrics%22%3A%5B%7B%22id%22%3A%22mean%22%2C%22label%22%3A%22Average%22%2C%22chartLabel%22%3A%22Average%22%2C%22themeColor%22%3A%22infographicB%22%7D%2C%7B%22id%22%3A%22std%22%2C%22label%22%3A%22St+Deviation%22%2C%22chartLabel%22%3A%22St+Deviation%22%2C%22themeColor%22%3A%22infographicD%22%7D%5D%2C%22colorMap%22%3A%22blues%22%2C%22scale%22%3A%7B%22min%22%3A0%2C%22max%22%3A0.3%7D%7D%7D%2C%7B%22id%22%3A%22casa-gfed-co2-flux%22%2C%22settings%22%3A%7B%22isVisible%22%3Atrue%2C%22opacity%22%3A100%2C%22analysisMetrics%22%3A%5B%7B%22id%22%3A%22mean%22%2C%22label%22%3A%22Average%22%2C%22chartLabel%22%3A%22Average%22%2C%22themeColor%22%3A%22infographicB%22%7D%2C%7B%22id%22%3A%22std%22%2C%22label%22%3A%22St+Deviation%22%2C%22chartLabel%22%3A%22St+Deviation%22%2C%22themeColor%22%3A%22infographicD%22%7D%5D%7D%7D%5D
-export function ExplorationAndAnalysisCompound({
-  selectedDay,
-  setSelectedDay,
-  selectedCompareDay,
-  setSelectedCompareDay,
-  setDatasets,
-  datasets,
-  children
-}: {
-  children: React.ReactNode;
-} & EACompoundContextType) {
-  return (
-    <EACompoundContext.Provider
-      value={{
-        selectedDay,
-        setSelectedDay,
-        selectedCompareDay,
-        setSelectedCompareDay,
-        setDatasets,
-        datasets
-      }}
-    >
-      {children}
-    </EACompoundContext.Provider>
-  );
-}
-
-function EAMapCompound(props) {
-  const { datasets, selectedDay, setDatasets, selectedCompareDay } =
-    useEACompoundContext();
-  return (
-    <ExplorationMap
-      selectedDay={selectedDay}
-      selectedCompareDay={selectedCompareDay}
-      setDatasets={setDatasets}
-      datasets={datasets}
-      {...props}
-    />
-  );
-}
-
-function EATimelineCompound(props) {
-  const {
-    selectedDay,
-    setSelectedDay,
-    selectedCompareDay,
-    setSelectedCompareDay,
-    datasets
-  } = useEACompoundContext();
-  return (
-    <Timeline
-      selectedDay={selectedDay}
-      setSelectedDay={setSelectedDay}
-      selectedCompareDay={selectedCompareDay}
-      setSelectedCompareDay={setSelectedCompareDay}
-      datasets={datasets}
-      {...props}
-    />
-  );
-}
-
-// This is the DataLayerCardWithSync
-// doesn't need EAContextHere so it can be used outside of the ExplorationAndAnalysisCompound
-export function EADataLayerCard(props: {
-  dataset: TimelineDataset;
-  setLayerInfo: (data: any) => void;
-}) {
-  const { dataset, setLayerInfo } = props;
-
-  const [isVisible, setVisible] = useTimelineDatasetVisibility(dataset.data.id);
-  const [colorMap, setColorMap] = useTimelineDatasetColormap(dataset.data.id);
-  const [colorMapScale, setColorMapScale] = useTimelineDatasetColormapScale(
-    dataset.data.id
-  );
-
-  const onClickLayerInfo = useCallback(() => {
-    const data: any = {
-      name: dataset.data.name,
-      description: dataset.data.description,
-      info: dataset.data.info,
-      parentData: dataset.data.parentDataset
-    };
-    setLayerInfo(data);
-  }, [dataset, setLayerInfo]);
-
-  return (
-    <DataLayerCard
-      dataset={dataset}
-      colorMap={colorMap}
-      colorMapScale={colorMapScale}
-      setColorMap={setColorMap}
-      setColorMapScale={setColorMapScale}
-      isVisible={isVisible}
-      setVisible={setVisible}
-      onClickLayerInfo={onClickLayerInfo}
-    />
-  );
-}
-
-ExplorationAndAnalysisCompound.Map = EAMapCompound;
-ExplorationAndAnalysisCompound.Timeline = EATimelineCompound;
-ExplorationAndAnalysisCompound.DataLayerCard = EADataLayerCard;
